[SIG-1844] move componenents in containers
